# Conflicts:
#	src/service/Project/src/ProjectWebApi/Controllers/FileImportV4Controller.cs
#	src/service/Project/src/ProjectWebApi/Properties/launchSettings.json

+++ src/service/TRex/src/netstandard/VSS.TRex.Designs/GridFabric/Arguments/CalculateDesignElevationPatchArgument.cs
@@ -12,12 +12,12 @@ namespace VSS.TRex.Designs.GridFabric.Arguments
     /// <summary>
     /// The X origin location for the patch of elevations, or spot elevation, to be computed from
     /// </summary>
-    public uint OriginX { get; set; }
+    public int OriginX { get; set; }
 
     /// <summary>
     /// The Y origin location for the patch of elevations, or spot elevation, to be computed from
     /// </summary>
-    public uint OriginY { get; set; }
+    public int OriginY { get; set; }
 
     /// <summary>
     /// The cell stepping size to move between points in the patch being interpolated
@@ -45,8 +45,8 @@ namespace VSS.TRex.Designs.GridFabric.Arguments
     /// <param name="cellSize"></param>
     /// <param name="referenceDesign"></param>
     public CalculateDesignElevationPatchArgument(Guid siteModelID,
-      uint originX,
-      uint originY,
+      int originX,
+      int originY,
       double cellSize,
       DesignOffset referenceDesign) : base(siteModelID, referenceDesign)
     {
@@ -74,8 +74,8 @@ namespace VSS.TRex.Designs.GridFabric.Arguments
 
       VersionSerializationHelper.EmitVersionByte(writer, VERSION_NUMBER);
 
-      writer.WriteInt((int)OriginX);
-      writer.WriteInt((int)OriginY);
+      writer.WriteInt(OriginX);
+      writer.WriteInt(OriginY);
       writer.WriteDouble(CellSize);
     }
 
@@ -89,8 +89,8 @@ namespace VSS.TRex.Designs.GridFabric.Arguments
 
       VersionSerializationHelper.CheckVersionByte(reader, VERSION_NUMBER);
 
-      OriginX = (uint)reader.ReadInt();
-      OriginY = (uint)reader.ReadInt();
+      OriginX = reader.ReadInt();
+      OriginY = reader.ReadInt();
       CellSize = reader.ReadDouble();
     }
   }
